Extracted Create books from app
